converted routes to user friendly version
