+++ app/assets/stylesheets/group.css
@@ -85,6 +85,7 @@ list-style-type: none;
   }
 
 
+
 #sortable {
   list-style-type: none;
   margin-left: auto;
