+++ app/helpers/groups_helper.rb
@@ -1,0 +1,2 @@
+module GroupsHelper
+end
