--- app/views/courses/new.html.erb
@@ -1,7 +1,0 @@
-
-<%= form_for 'users[courses]', url: {action: "create"} do |f| %>
-  <%= label_tag("course_title", "Course Title: ")%>
-  <%= text_field_tag("course_title") %>
-
-  <%= f.submit "Create New Course" %>
-<% end %>
