+++ app/views/user_courses/index.html.erb
@@ -1,0 +1,25 @@
+<h1>Welcome to SmartGroup!</h1>
+
+<h2 class="subheader">Add A Course</h2>
+
+  <%= link_to("Upload Class", parser_path, class: "button success") %>
+
+  <%= link_to("Enter Class", new_user_course_path, class: "button success") %>
+<h2 class="subheader">Current Courses</h2>
+
+<table id="course_list">
+  <% @courses.reverse.each do |course| %>
+    <tr>
+      <td ><b><%= course.name %></b></td>
+
+    <!-- HREF TO GROUPS MAKE INSTEAD -->
+      <td><%= link_to("Group Generator", new_course_group_path(course), class: "small button", id: "opt_bttn") %></td>
+      <td><%= link_to("Metics / Graphs", "#", class: "small button", id: "opt_bttn") %></td>
+      <td><%= link_to("Edit Roster", "#", class: "small button", id: "opt_bttn") %></td>
+    </tr>
+  <% end %>
+</table>
+
+
+
+
